controllers/dispalyTextController.js: タイピングのお題データを一旦切り分け
