+++ src/controllers/dispalyTextController.js
@@ -1,0 +1,15 @@
+// 後々はDBから取得したデータをjs配列に変換したりする場所にする想定
+
+const textDataArr = [
+  {
+    displayTextKanji: "有り難う",
+    dispalyTextKana: "ありがとう"
+  },
+  {
+    displayTextKanji: "素晴らしい",
+    dispalyTextKana: "すばらしい"
+  }
+];
+
+
+export default textDataArr;
